+++ src/vsa/scientific/__init__.py
@@ -1,0 +1,1 @@
+"""Scientific credibility policies."""